Move event helper function into server since it's only used there
