+++ src/desktop/server.rs
@@ -9,12 +9,19 @@ use piston_window::{
     WindowSettings,
 };
 
+use input::{
+    Event as PistonEvent,
+    Input, ButtonArgs,
+    ButtonState,
+    Button,
+    Motion,
+};
+
 use super::messenger;
 use app::{ReadOnly, TurtleApp};
 use renderer::Renderer;
 use query::{Query, DrawingCommand, Request, StateUpdate, Response};
 use {Event};
-use event::from_piston_event;
 use extensions::ConvertScreenCoordinates;
 
 /// If this process is the child rendering process, enter the rendering loop and do not pass
@@ -210,4 +217,48 @@ fn run_render_loop(drawing_rx: mpsc::Receiver<DrawingCommand>,
             renderer.render(c, g, center, &drawing, &temporary_path, &turtle);
         });
     }
+}
+
+/// Attempts to convert a piston Event to our event type
+fn from_piston_event<F>(event: &PistonEvent, to_local_coords: F) -> Option<Event>
+    where F: FnOnce(::Point) -> ::Point {
+    use self::Event::*;
+
+    let input_event = match *event {
+        PistonEvent::Input(ref input_event) => input_event,
+        _ => return None,
+    };
+
+    Some(match *input_event {
+        Input::Button(ButtonArgs {state, button, scancode: _}) => match state {
+            ButtonState::Press => match button {
+                Button::Keyboard(key) => KeyPressed(key),
+                Button::Mouse(button) => MouseButtonPressed(button),
+                Button::Controller(button) => ControllerButtonPressed(button),
+            },
+            ButtonState::Release => match button {
+                Button::Keyboard(key) => KeyReleased(key),
+                Button::Mouse(button) => MouseButtonReleased(button),
+                Button::Controller(button) => ControllerButtonReleased(button),
+            },
+        },
+        Input::Move(motion) => match motion {
+            Motion::MouseCursor(x, y) => {
+                let local = to_local_coords([x, y]);
+                MouseMove {x: local[0], y: local[1]}
+            },
+            // Ignored in favor of MouseCursor
+            Motion::MouseRelative(..) => return None,
+            Motion::MouseScroll(x, y) => MouseScroll {x, y},
+            Motion::ControllerAxis(axis) => ControllerAxisChange(axis),
+            Motion::Touch(touch) => Touch(touch),
+        },
+        // Ignored because this value doesn't produce text reliably for all keys
+        // (especially when ctrl is pressed)
+        Input::Text(_) => return None,
+        Input::Resize(x, y) => WindowResized {x, y},
+        Input::Focus(focused) => WindowFocused(focused),
+        Input::Cursor(cursor) => WindowCursor(cursor),
+        Input::Close(_) => WindowClosed,
+    })
 }
